heap and merge sort

+++ sort.h
@@ -28,10 +28,10 @@ void quick_sort(int *array, int size);
 void shell_sort(int *array, size_t size);
 void cocktail_sort_list(listint_t **list);
 void counting_sort(int *array, size_t size);
-void merge_sort(int *array, int size);
-void heap_sort(int *array, int size);
-void radix_sort(int *array, int size);
-void bitonic_sort(int *array, int size);
-void quick_sort_hoare(int *array, int size);
+void merge_sort(int *array, size_t size);
+void heap_sort(int *array, size_t size);
+void radix_sort(int *array, size_t size);
+void bitonic_sort(int *array, size_t size);
+void quick_sort_hoare(int *array, size_t size);
 
 #endif /*SORT_H*/
